Fix string replacement targets in bayesian table header
The targeted table headers in the dataframe are of the form 'p_equal',
without a backslash. The backslash would only be necessary for LaTeX
export, to escape the underscore.

The wrong replacement left the table headers intact, causing issues in
rendering the table LaTeX because of the underscores mentioned above.

This fix re-establishes the indented behaviour of pretty-printing the
probabilities.